Found a bunch of bugs in the server side stuff, so am refocusing on that.  Blackberry implementation is coming along, but I suspect both Android and Blackberry and broken, at the moment.

+++ server-side/openmeap-admin-web/src/com/openmeap/admin/web/backing/AddModifyApplicationBacking.java
@@ -33,6 +33,7 @@ import java.util.Map;
 
 import javax.persistence.PersistenceException;
 
+import org.apache.commons.lang.exception.ExceptionUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -116,45 +117,58 @@ public class AddModifyApplicationBacking extends AbstractTemplatedSectionBacking
 			
 			app = createApplicationFromParameters(app,parameterMap, events);
 			
-			if( events.size()==0 ) {
-				try {
-					app.setLastModifier(firstValue("userPrincipalName",parameterMap));
-					modelManager.begin();
-					app = modelManager.addModify(app,events);
-					modelManager.commit(events);
-					events.add( new MessagesEvent("Application successfully created/modified!") );
-				} catch( InvalidPropertiesException ipe ) {
-					modelManager.rollback();
-					events.add( new MessagesEvent(ipe.getMessage()) );
-				} catch( PersistenceException pe ) {
-					modelManager.rollback();
-					events.add( new MessagesEvent(pe.getMessage()) );							
+			if( ParameterMapUtils.firstValue("submit",parameterMap).equals("true") ) {
+				
+				if( events.size()==0 ) {
+					try {
+						app.setLastModifier(firstValue("userPrincipalName",parameterMap));
+						modelManager.begin();
+						app = modelManager.addModify(app,events);
+						modelManager.commit(events);
+						events.add( new MessagesEvent("Application successfully created/modified!") );
+					} catch( InvalidPropertiesException e ) {
+						events.add(new MessagesEvent(String.format("Application add/modify failed: %s %s"
+								,ExceptionUtils.getRootCauseMessage(e)
+								,ExceptionUtils.getRootCauseStackTrace(e)[0])));
+						logger.error("Add/Modify application with id "+app.getId()+" failed",e);
+						modelManager.rollback();
+					} catch( PersistenceException e ) {
+						events.add(new MessagesEvent(String.format("Application add/modify failed: %s %s"
+								,ExceptionUtils.getRootCauseMessage(e)
+								,ExceptionUtils.getRootCauseStackTrace(e)[0])));
+						logger.error("Add/Modify application with id "+app.getId()+" failed",e);
+						modelManager.rollback();							
+					}
 				}
+				
+				if( app==null && ParameterMapUtils.notEmpty(FormConstants.APP_ID,parameterMap) )
+					app = modelManager.getModelService().findByPrimaryKey(Application.class, 
+							Long.valueOf(ParameterMapUtils.firstValue(FormConstants.APP_ID, parameterMap)));
+				
 			}
 			
-			if( app==null && ParameterMapUtils.notEmpty(FormConstants.APP_ID,parameterMap) )
-				app = modelManager.getModelService().findByPrimaryKey(Application.class, 
-						Long.valueOf(ParameterMapUtils.firstValue(FormConstants.APP_ID, parameterMap)));
-			
-			if( ParameterMapUtils.notEmpty(FormConstants.DELETE,parameterMap) && ParameterMapUtils.notEmpty("deleteConfirm",parameterMap) ) {
+			if( ParameterMapUtils.firstValue("delete",parameterMap).equals("true") ) {
 				
-					if( ParameterMapUtils.firstValue("deleteConfirm", parameterMap).equals(FormConstants.APP_DELETE_CONFIRM_TEXT) ) {
-						try {
-							modelManager.begin();
-							modelManager.delete(app,events);
-							modelManager.commit(events);
-							events.add( new MessagesEvent("Application successfully deleted!") );
-							app = null;
-							// we remove the applicationId parameter, so that the form can populate empty
-							parameterMap.remove(FormConstants.APP_ID);
-						} catch(Exception e) {
-							events.add( new MessagesEvent("Application delete failed!") );
-							logger.error("Deleting application with id "+app.getId()+" failed",e);
-							modelManager.rollback();
-						}
-					} else {
-						
-						events.add( new MessagesEvent("You must confirm your desire to delete by typing in the delete confirmation message.") );
+				if( !ParameterMapUtils.empty("deleteConfirm", parameterMap)
+						&& ParameterMapUtils.firstValue("deleteConfirm", parameterMap).equals(FormConstants.APP_DELETE_CONFIRM_TEXT) ) {
+					try {
+						modelManager.begin();
+						modelManager.delete(app,events);
+						modelManager.commit(events);
+						events.add( new MessagesEvent("Application successfully deleted!") );
+						app = null;
+						// we remove the applicationId parameter, so that the form can populate empty
+						parameterMap.remove(FormConstants.APP_ID);
+					} catch(Exception e) {
+						events.add(new MessagesEvent(String.format("Application delete failed: %s %s"
+								,ExceptionUtils.getRootCauseMessage(e)
+								,ExceptionUtils.getRootCauseStackTrace(e)[0])));
+						logger.error("Deleting application with id "+app.getId()+" failed",e);
+						modelManager.rollback();
+					}
+				} else {
+					
+					events.add( new MessagesEvent("You must confirm your desire to delete by typing in the delete confirmation message.") );
 				}
 			} 
 		}
